Migrate internal storage codec from rkyv to flatbuffers
